+++ packages/core/lib/combo/custom/custom-objects-builder.ts
@@ -337,6 +337,46 @@ export class CustomObjectsBuilder {
     return { name: 'EQ_HOOKSHOT', ...editor.build() };
   }
 
+  private async makeEqSlingshot(): Promise<CustomObject> {
+    const editor = new ObjectEditor(0xa);
+    const obj = await this.getFile('oot', 'objects/object_link_child');
+    editor.loadSegment(0x06, obj);
+
+    const dataFPAddr = 0x06018048;
+    let dataFP = editor.listData(dataFPAddr)!;
+    dataFP = editor.stripList(dataFP, 0x06018128 - dataFPAddr, 0x06018470 - dataFPAddr);
+    editor.submitList(dataFP);
+
+    editor.submitList(editor.listData(0x060221a8)!); /* First person string */
+
+    const dataTPAddr = 0x06015df0;
+    let dataTP = editor.listData(dataFPAddr)!;
+    dataTP = editor.stripList(dataTP, 0x06015F18 - dataTPAddr, 0x06015FC8 - dataTPAddr);
+    editor.submitList(dataTP);
+
+    return { name: 'EQ_SLINGSHOT', ...editor.build() };
+  }
+
+  private async makeEqBow(): Promise<CustomObject> {
+    const editor = new ObjectEditor(0xa);
+    const obj = await this.getFile('oot', 'objects/object_link_boy');
+    editor.loadSegment(0x06, obj);
+
+    const dataFPAddr = 0x0602a248;
+    let dataFP = editor.listData(dataFPAddr)!;
+    dataFP = editor.stripList(dataFP, 0x0602A3F8 - dataFPAddr, 0x0602a730 - dataFPAddr);
+    editor.submitList(dataFP);
+
+    editor.submitList(editor.listData(0x0602b108)!); /* First person string */
+
+    const dataTPAddr = 0x06022da8;
+    let dataTP = editor.listData(dataTPAddr)!;
+    dataTP = editor.stripList(dataTP, 0x06022f00 - dataTPAddr, 0x06023158 - dataTPAddr);
+    editor.submitList(dataTP);
+
+    return { name: 'EQ_BOW', ...editor.build() };
+  }
+
   async build(): Promise<CustomObject[]> {
     return [
       await this.makeEqKokiriSword(),
@@ -361,6 +401,8 @@ export class CustomObjectsBuilder {
       await this.makeEqOcarinaFairy(),
       await this.makeEqBoomerang(),
       await this.makeEqHookshot(),
+      await this.makeEqSlingshot(),
+      await this.makeEqBow(),
       //await this.simpleExtract('LIMB_OOT_CHILD_LHAND_CLOSED', 'oot', 'objects/object_link_child', [], 0x06, 0x0a),
     ];
   }