+++ packages/core/lib/combo/patch-build/index.ts
@@ -51,6 +51,8 @@ function asmPatchGroups(world: World, settings: Settings) {
     OOT_AGELESS_BOOMERANG: settings.agelessBoomerang,
     OOT_AGELESS_HAMMER: settings.agelessHammer,
     OOT_AGELESS_HOOKSHOT: settings.agelessHookshot,
+    OOT_AGELESS_SLINGSHOT: settings.agelessSlingshot,
+    OOT_AGELESS_BOW: settings.agelessBow,
     OOT_AGELESS_CHILD_TRADE: settings.agelessChildTrade,
     OOT_BLUE_FIRE_ARROWS: settings.blueFireArrows,
     OOT_SUNLIGHT_ARROWS: settings.sunlightArrows,